simplify hydro remix : moving classes declarations for remixing storage to header files
